@jvsena42 2025-09-04
### Benchmark

Before:
```
2025-09-04 14:35:38.177  9188-9251  APP                     to.bitkit.dev                        V  PERF: ActivityService.syncLdkNodePayments took 0.173 seconds on core-queue [ServiceQueue.kt:74]
2025-09-04 14:35:46.720  9188-9316  APP                     to.bitkit.dev                        V  PERF: LightningService.sync took 7.806 seconds on ldk-queue [ServiceQueue.kt:74]
```

After:
```
2025-09-04 14:43:36.041  9923-9975  APP                     to.bitkit.dev                        I  INFOℹ️: Synced 145 payments successfully, 0 failed - CoreService [CoreService.kt:279]
2025-09-04 14:43:36.043  9923-9984  APP                     to.bitkit.dev                        V  PERF: ActivityService.syncLdkNodePayments took 0.082 seconds on core-queue [ServiceQueue.kt:74]

2025-09-04 14:43:44.829  9923-10057 APP                     to.bitkit.dev                        V  PERF: LightningService.sync took 7.836 seconds on ldk-queue [ServiceQueue.kt:74]
```

Got some improvement, but the real bottleneck is LightningService.sync